+++ integration_tests/scripts/core_integration_starter.js
@@ -1,3 +1,6 @@
+/*
+ * Copyright (C) 2022-present The Kraken authors. All rights reserved.
+ */
 const { spawn, spawnSync } = require('child_process');
 const path = require('path');