+++ kraken/lib/src/bridge/to_native.dart
@@ -1,6 +1,5 @@
 /*
- * Copyright (C) 2021-present Alibaba Inc. All rights reserved.
- * Author: Kraken Team.
+ * Copyright (C) 2021-present The Kraken authors. All rights reserved.
  */
 
 import 'dart:async';
@@ -118,7 +117,7 @@ void emitUIEvent(
   Pointer<Void> rawEvent = event.toRaw().cast<Void>();
   bool isCustomEvent = event is CustomEvent;
   Pointer<NativeString> eventTypeString = stringToNativeString(event.type);
-  // @TODO: Make Event inhert BindingObject to pass value from bridge to dart.
+  // @TODO: Make Event inherit BindingObject to pass value from bridge to dart.
   int propagationStopped = dispatchEvent(contextId, nativeBindingObject, eventTypeString, rawEvent, isCustomEvent ? 1 : 0);
   event.propagationStopped = propagationStopped == 1 ? true : false;
   freeNativeString(eventTypeString);