+++ kraken/lib/src/css/style_declaration.dart
@@ -1,6 +1,5 @@
 /*
- * Copyright (C) 2019-present Alibaba Inc. All rights reserved.
- * Author: Kraken Team.
+ * Copyright (C) 2019-present The Kraken authors. All rights reserved.
  */
 
 import 'package:kraken/css.dart';
@@ -495,7 +494,7 @@ class CSSStyleDeclaration {
 
   void _emitPropertyChanged(String property, String? original, String present) {
     if (original == present) return;
-    
+
     if (onStyleChanged != null) {
       onStyleChanged!(property, original, present);
     }