+++ kraken/lib/src/css/values/function.dart
@@ -1,6 +1,5 @@
 /*
- * Copyright (C) 2021-present Alibaba Inc. All rights reserved.
- * Author: Kraken Team.
+ * Copyright (C) 2021-present The Kraken authors. All rights reserved.
  */
 
 // CSS Values and Units: https://drafts.csswg.org/css-values-3/#functional-notations