+++ kraken/lib/src/module/module_manager.dart
@@ -1,3 +1,6 @@
+/*
+ * Copyright (C) 2021-present The Kraken authors. All rights reserved.
+ */
 import 'dart:convert';
 
 import 'package:kraken/bridge.dart' as bridge;